requirements.txt
pytest
pyyaml
black
ruff
typer
rich